Split rule definitions and associated unit tests into separate files.
