+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@justeat/f-validate",
-  "description": "",
-  "version": "0.6.0",
+  "description": "Fozzie vanilla JS Validation Component",
+  "version": "0.6.1",
   "main": "dist/index.js",
   "homepage": "https://github.com/justeat/f-validate",
   "contributors": [
