+++ src/rules/index.js
@@ -1,0 +1,17 @@
+import required from './required';
+import maxlength from './maxlength';
+import minlength from './minlength';
+import pattern from './pattern';
+import email from './email';
+import matches from './matches';
+import custom from './custom';
+
+export default {
+    required,
+    maxlength,
+    minlength,
+    pattern,
+    email,
+    matches,
+    custom
+};
